requirements.txt
cryptography
qrcode
fpdf
Pillow
